feat: 수정 로직 분리 - 제출 로직 역할에 따라 분리 - 화면 분리
